+++ shopify/packages/ajax-cart-components/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @brytdesigns/shopify-ajax-cart-components
 
+## 1.1.1
+
+### Patch Changes
+
+- @brytdesigns/shopify-ajax-cart@1.0.6
+
 ## 1.1.0
 
 ### Minor Changes